Breaking(!) EnvironmentManager update
This update involves a rework of the EnvironmentManager which makes it
incompatible with older versions of the MATE-Server.

The EnvironmentManager is now communicating with the MATE-Server using
the Message class.

The EnvironmentManager now maintains one connection instead of reopening
a new connection every time a new command is being send to the server.

The EnvironmentManager is now dynamically initialized and maintained in
the Registry class.

Fixes #6